tests/common: move device-related fixtures to 'common'
preauth tests will be done in a separate module, and will
reuse these.

changelog: none
