+++ tests/tests/test_device.py
@@ -5,45 +5,17 @@ import bravado
 import pytest
 
 from common import Device, DevAuthorizer, \
-    device_auth_req, \
+    device_auth_req, make_devices, devices, \
     clean_migrated_db, clean_db, mongo, cli, \
     management_api, internal_api, device_api, \
     tenant_foobar
+
 
 import mockserver
 import deviceadm
 import inventory
 
 
-def make_devices(device_api, devcount=1, tenant_token=""):
-    print('device count to generate', devcount)
-
-    url = device_api.auth_requests_url
-
-    out_devices = []
-    with deviceadm.run_fake_for_device(deviceadm.ANY_DEVICE) as server:
-        for _ in range(devcount):
-            dev = Device()
-            da = DevAuthorizer(tenant_token=tenant_token)
-            # poke devauth so that device appears
-            rsp = device_auth_req(url, da, dev)
-            assert rsp.status_code == 401
-            out_devices.append((dev, da))
-
-    return out_devices
-
-
-@pytest.yield_fixture(scope='function')
-def devices(device_api, management_api, clean_migrated_db, request):
-    """Make unauthorized devices. The fixture can be parametrized a number of
-    devices to make. Yields a list of tuples:
-    (instance of Device, instance of DevAuthorizer)"""
-    if not hasattr(request, 'param'):
-        devcount = 1
-    else:
-        devcount = int(request.param)
-
-    yield make_devices(device_api, devcount)
 
 
 @pytest.yield_fixture(scope='function')
